Split database specific install/setup code to their own class
